### 2022-01-30 #4106 @arfon
:wave: @Kevin-Mattheus-Moerman @prashjha – I know you both have pretty full (editorial) plates right now. Are there any other JOSS editors you think could be a good fit here?
