### 2022-02-16 #4106 @Kevin-Mattheus-Moerman
@urban-fasel thanks for this submission. I am currently looking for reviewers. If you can suggest any reviewers that would be helpful too. You can mention their github handle (without the @ symbol) here.

Could you clarify if this submission is compatible with Octave? Or does it require special MATLAB functions/toolboxes that are note part of Octave. I ask this as some reviewers can only review Octave rather than MATLAB submissions. Thanks.
