### 2022-02-16 #4106 @urban-fasel
Hi @Kevin-Mattheus-Moerman, thanks for being the editor of our submission!
Some potential reviewers from the joss_reviewers list: vaishnavtv, wahls, mbudisic, balaje
and some additional potential reviewers: nairaditya, ben-herrmann, dylaneij, rschmehl, marconij
The submission is not compatible with Octave as it uses some functions that are not implemented in Octave yet.
