### 2022-02-17 #4106 @Kevin-Mattheus-Moerman
@weymouth I was wondering if this submission on flexible wings featuring FSI is of interest to you and if you could review it for [JOSS](https://joss.theoj.org/). It does require MATLAB though. If you are not available/able, perhaps you could suggest a colleague? Thanks.

Short paper: https://github.com/openjournals/joss-papers/blob/joss.04106/joss.04106/10.21105.joss.04106.pdf
Software: https://github.com/urban-fasel/FlexWingROM
