### 2022-02-22 #4106 @Kevin-Mattheus-Moerman
@oriolcg this is a prereview issue where we assign reviewers. Once we have at least two reviewers I'll trigger a review issue. In the review issue you'll have a checklist to go through which will guide you through the review process. Here is an example of a finished review issue:
https://github.com/openjournals/joss-reviews/issues/3053

The process should be fairly smooth and self-explanatory, but if you want to learn more you can read all about it here too:
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Let me know if you have any questions. Thanks for your help!
