### 2022-02-22 #4106 @Kevin-Mattheus-Moerman
@chennachaos I feel like I'm close to "over inviting" you to review for JOSS, but do let me know if this work is your cup of tea, and would, despite the reliance on MATLAB, be interesting to review.
