### 2022-02-22 #4106 @Kevin-Mattheus-Moerman
@Nicola-Fonzi @dciliberti @Philipp-Conen @JaroslavHron @albansouche would you be interested in reviewing this submission for [the Journal of Open Source Software (JOSS)](https://joss.theoj.org/). The work is entitled: _A matlab framework for data-driven reduced-order modeling of flexible wings_.

Note this submission is not Octave compatible so reviewing it does require a MATLAB installation.

JOSS reviews take place on GitHub and focus on the review and evaluation of [the software](https://github.com/urban-fasel/FlexWingROM) as well as a [very short paper](https://github.com/openjournals/joss-papers/blob/joss.04106/joss.04106/10.21105.joss.04106.pdf).

This is a pre-review issue where we find and assign reviewers. If you are interested I'll add you as a reviewer and once we have sufficient reviewers I'll trigger a dedicated review issue (see this example of a finished review issue: https://github.com/openjournals/joss-reviews/issues/2089).

Thanks.
