### 2022-02-22 #4106 @Kevin-Mattheus-Moerman
@Nicola-Fonzi my sincere apologies. I see you are on the paper, sorry I did not see your name as a contributor for the software (https://github.com/urban-fasel/FlexWingROM/graphs/contributors) so accidentally invited you.
