### 2022-02-23 #4106 @dciliberti
Good evening,sorry I am not available. Best regards,Danilo Ciliberti Da: Kevin Mattheus MoermanInviato: martedì 22 febbraio 2022 14:49A: openjournals/joss-reviewsCc: Danilo Ciliberti; MentionOggetto: Re: [openjournals/joss-reviews] [PRE REVIEW]: FlexWing-ROM: A matlab framework for data-driven reduced-order modeling of flexible wings (Issue ***@***.*** @dciliberti @Philipp-Conen @JaroslavHron @albansouche would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS). The work is entitled: A matlab framework for data-driven reduced-order modeling of flexible wings.Note this submission is not Octave compatible so reviewing it does require a MATLAB installation.JOSS reviews take place on GitHub and focus on the review and evaluation of the software as well as a very short paper.This is a pre-review issue where we find and assign reviewers. If you are interested I'll add you as a reviewer and once we have sufficient reviewers I'll trigger a dedicated review issue (see this example of a finished review issue: #2089).Thanks.—Reply to this email directly, view it on GitHub, or unsubscribe.You are receiving this because you were mentioned.Message ID: ***@***.***>
