### 2022-02-28 #4211 @editorialbot
[REVIEW]: FlexWing-ROM: A matlab framework for data-driven reduced-order modeling of flexible wings
**Submitting author:** @urban-fasel (<a href="http://orcid.org/0000-0002-3716-686X">Urban Fasel</a>)
**Repository:** https://github.com/urban-fasel/FlexWingROM
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
