### 2022-05-13 #4211 @omtazi
Hello @Kevin-Mattheus-Moerman @oriolcg , I'm having trouble running the example1 and example2 codes. I get the following message when I run example2_NACA6418_unsteadyFSI.m . Do you have this too ?

`Unable to resolve the name 'parallel.Settings'.

Error in Assembly/adjust_parallel_settings (line 26)
            ps = parallel.Settings;

Error in Assembly/set.parallelized (line 22)
            adjust_parallel_settings(self);

Error in Assembly (line 16)
                self.parallelized = false;

Error in runYetAnotherFEcode (line 147)
PlateAssembly = Assembly(myMesh);

Error in generateWingModelStructure (line 1374)
[M,K,wP] = runYetAnotherFEcode(wingProperties, wingDesign);

Error in example2_NACA6418_unsteadyFSI (line 46)
    wingModelStructure = generateWingModelStructure(wingDesign,simParam);`
