### 2022-06-01 #4211 @urban-fasel
Hi @Kevin-Mattheus-Moerman, yes! And sorry for not directly replying to the issue.

Hi @omtazi, the issue is related to the "yetAnotherFEcode" that we use to build the FE model of the wing. This package uses the "Parallel Computing Toolbox". If you install the toolbox, you should be able to run the code.

In case other toolboxes are missing, the required toolboxes for e.g. example2 can be found and printed in the command line like this:
[~,pList] = matlab.codetools.requiredFilesAndProducts('example2_NACA6418_unsteadyFSI.m');
{pList.Name}'

I hope this helps and please let me know if you still get the error message.
