### 2022-06-25 #4211 @oriolcg
@urban-fasel , I get the following error when executing the code:
```'ss' requires one of the following:
  Control System Toolbox
  DSP System Toolbox
  Model Predictive Control Toolbox
  Signal Processing Toolbox

Error in generate_gramian (line 19)
            sys = ss(A_aIODMDc,B_aIODMDc,C_aIODMDc,D_aIODMDc,0.006);

Error in Obtain_BMD_model (line 67)
    [W_o, W_c] = generate_gramian(Snapshot(:,:,k), Snapshot_input(:,:,k), V0, type_string_gram, loadGramians, rankMax, simInput);

Error in ROMsim (line 33)
        ROMs = Obtain_BMD_model(Snapshot, Snapshot_input, V0, type_string, rankROM.rankMaxBMD, simInput);

Error in ROM (line 119)
ROMsim(V0,rankROM,type_string,dT,iTest,simInput,simOutFULL);

Error in MAIN (line 187)
    ROM(paramFSI);
```
Please, state in the README the required toolbox that should be installed before executing
