### 2022-06-28 #4211 @urban-fasel
@oriolcg and @Kevin-Mattheus-Moerman  thanks for the comments, I added the toolbox dependencies to the readme:

### Dependencies

The following matlab toolboxes are required to run the code:

* Control System Toolbox
* System Identification Toolbox
* Statistics and Machine Learning Toolbox
* Parallel Computing Toolbox
