### 2022-07-06 #4211 @omtazi
@Kevin-Mattheus-Moerman I checked most of the items of the checklist. However, I can't find the "Community guidelines" part of the documentation for people wishing to contribute to the software.
