### 2022-07-06 #4211 @urban-fasel
@omtazi I forgot to add the community guidelines, I just added them to the readme!
