### 2022-07-26 #4211 @omtazi
Hello, I wrote a synthesis of the comments I had on the code and the article. Let me know if you need anything else.

FlexWingROM is a useful Matlab framework that allows to model flexible wings. The paper clearly gives context on the need that the code aims at responding and that motivates this work, and pedagogically explains the approach adopted to contribute to the field of aeroelastic reduce-order modeling. The code is split in different parts that can be used independently: it allows the user to model a flexible wing and solve the corresponding FSI problem using a parameterized wing generator and a finite element solver. Reduced-order models can be built using snapshots of the simulation using three different ROM methods. The code allows the user to compare the results obtained with the ROM to the FE results to evaluate the accuracy of the model; it also allows to compare the performance of the three ROM methods to each other.
	The documentation is clear and well written. The installation goes smoothly as the necessary toolboxes are stated. The paper and the GitHub instructions are very helpful in understanding the approach of this work and the theory behind the code. The implementation is pretty explicit, and the comments written in the code allow to follow the reasoning of the authors. It seems to me that, given enough time, one can understand each step of the implementation using the comments in the code and the documentation. Examples are given so the user can try by themselves the different features of the framework and play with the main parameters.
	One possible improvement that comes to my mind would be to explicitly mention in the documentation all the physical quantities that are represented in the output graphs, as their symbol might not always be sufficient to unambiguously identify them.
	To sum up, FlexWing-ROM is an interesting tool to model flexible wings using ROMs that could prove to be useful for students and researchers and deserves to be published JOSS.
