### 2022-12-02 urban-fasel/FlexWingROM#1 @Kevin-Mattheus-Moerman
Unrecognized function or variable 'dsegment'. Perhaps mex source file is missing? 
Upon running `unitTest_FlexWingROM.m` I get the following error:

```
Unrecognized function or variable 'dsegment'.

Error in dpoly (line 2)
    ds=dsegment(p,pv);
```

I cannot find this function with the source, although there is a file called `dsegment.mexw64`.
