### 2022-12-02 urban-fasel/FlexWingROM#2 @Kevin-Mattheus-Moerman
Clearly seperate original content from external libraries and conform to external software license terms
** IMPORTANT **
This code base contains codes obtained from [DistMesh](http://persson.berkeley.edu/distmesh/) and is currently not complying with its license terms:
> under the terms of the GNU General Public License as published by the Free Software Foundation; either version 2 of the License, or (at your option) any later version

- [ ] I propose you create a folder called `code_ext` and place within it any codes which is not your original work. For the `DistMesh` related components would create a folder called `DistMesh` in the `code_ext` folder. Then copy all required source code components into that folder and include a LICENSE file describing the license for `DistMesh`. I would do the same (dedicated folders with license files) for all other external packages/codes included. So, in short, clearly separate your codes from externally obtained codes. It is not permitted to share codes from others without complying to the license terms.

Note that the GPL license for `DistMesh` contains:
> But when you distribute the same sections as part of a whole which is a work based on the Program, the distribution of the whole must be on the terms of this License, whose permissions for other licensees extend to the entire whole, and thus to each and every part regardless of who wrote it.

Hence if you include `DistMesh` with `FlexWingRom` then:

- [ ]  `FlexWingRom` will also need to adopt a GPL license. I suggest [GPLv3](https://www.gnu.org/licenses/gpl-3.0.html), as long it is compatible.
