### 2022-12-02 urban-fasel/FlexWingROM#1 @Kevin-Mattheus-Moerman
I have now learned this relates to: https://github.com/urban-fasel/FlexWingROM/issues/2

I.e. insufficient code from `DistMesh` is included in the project. Please amend the codes as suggested, and perhaps include instructions for installation if any mex file compilation/processing steps are needed.
