### 2022-12-02 urban-fasel/FlexWingROM#2 @Kevin-Mattheus-Moerman
I have also found the following with its own license terms: https://uk.mathworks.com/matlabcentral/fileexchange/28337-cosspace

Please do the above (separate codes and adhere to license terms) for absolutely all types of external codes/libraries included.
