### 2022-12-02 #4211 @Kevin-Mattheus-Moerman
@urban-fasel I have opened some issues: https://github.com/urban-fasel/FlexWingROM/issues/1 and https://github.com/urban-fasel/FlexWingROM/issues/2, with some serious concerns. Please address these at your earliest convenience.
