### 2022-12-02 #4211 @Kevin-Mattheus-Moerman
@omtazi were you able to run `MAIN.m` and `unitTest_FlexWingROM.m`? Several functions were not included in the source or were included in an incomplete way (see also https://github.com/urban-fasel/FlexWingROM/issues/1).
