+++ code/YetAnotherFEcode-master/src/Element/MechanicalElements/Hex20Element.m
@@ -1,291 +1,0 @@
-classdef Hex20Element < Element
-    properties
-        nodes = []          % global coordinates of element nodes
-        nodeIDs = []        % the index location of element nodes
-        nDOFPerNode = 3     % number of DOFs per node
-        nNodes = 20         % number of nodes per element
-        nDim = 3            % number of dimensions in local coordinates
-    end
-    
-    properties
-        quadrature              % weights and points for gauss quadrature
-        Material                % Object of class Material
-        initialization          % some 0-matrices to speedup numerical integration
-    end
-    
-    properties (Dependent)
-        uniformBodyForce
-        vol                     % volume of the element
-    end
-    
-    methods
-        
-        % MINIMUM REQUIRED FUNCTIONS ______________________________________
-        
-        function self = Hex20Element(Material,Ngauss)
-            % _____________________________________________________________
-            %
-            % SELF-FUNCTION
-            % self = Hex20Element(Material,Ngauss)
-            % defines element's properties
-            %______________________________________________________________
-            if nargin == 1
-                self.Material = Material;
-                Ngauss = 2;
-                [x,w]=lgwt(Ngauss,-1,1);
-                self.quadrature.Ng = Ngauss;
-                self.quadrature.X = x;	% gauss integration points
-                self.quadrature.W = w;	% gauss integration weights
-            elseif nargin==2
-                self.Material = Material;
-                [x,w]=lgwt(Ngauss,-1,1);
-                self.quadrature.Ng = Ngauss;
-                self.quadrature.X = x;
-                self.quadrature.W = w;
-            end
-            % INIZIALIZATION of some matrices (this should speedup
-            % numerical integration)
-            v = self.Material.POISSONS_RATIO;
-            E = self.Material.YOUNGS_MODULUS;
-            C = [1-v v v 0 0 0; % isotropic material assumption
-                v 1-v v 0 0 0;
-                v v 1-v 0 0 0;
-                0 0 0 .5*(1-2*v) 0 0;
-                0 0 0 0 .5*(1-2*v) 0;
-                0 0 0 0 0 .5*(1-2*v)]*E/((1+v)*(1-2*v));
-            H = [1 0 0 0 0 0 0 0 0;
-                0 0 0 0 1 0 0 0 0;
-                0 0 0 0 0 0 0 0 1;
-                0 1 0 1 0 0 0 0 0;
-                0 0 1 0 0 0 1 0 0;
-                0 0 0 0 0 1 0 1 0];
-            self.initialization.A = zeros(6,9); % nonlinear strain
-            self.initialization.G = zeros(9,60);% shape function derivatives
-            self.initialization.Z = zeros(20);  % zero-matrix
-            self.initialization.K = zeros(60);  % stiffness-element matrix
-            self.initialization.F = zeros(60,1);% internal forces (element)
-            self.initialization.C = C;          % constitutive law matrix
-            self.initialization.H = H;          % linear strain
-        end
[…]
