+++ code/YetAnotherFEcode-master/src/TimeIntegration/GeneralizedAlpha.m
@@ -1,194 +1,0 @@
-classdef GeneralizedAlpha < handle
-%% Class for Generalized-alpha scheme for time Integration
-
-% We use the following reference in this implementation 
-% [1] Arnold, M., & Brüls, O. (2007). Convergence of the generalized-? scheme 
-% for constrained mechanical systems. Multibody System Dynamics, 18(2), 185–202. 
-% https://doi.org/10.1007/s11044-007-9084-0
-
-% For the time being, we consider systems without constraints.
-    properties
-        rho_inf = 0.7 % 0<spectral radius<1: rho_inf = 1 means undamped, rho_inf = 0 means asymptotic annhilation of high frequency response
-        alpha_m
-        alpha_f
-        beta
-        beta_p
-        h = 0 % time step size
-        gamma
-        gamma_p
-        tol = 1e-6
-        Solution
-        MaxNRit = 10
-        ATS = false     % where adaptive time stepping should be on (true) or not (false)
-        hmin = 0            % minimum timestep size
-        NROpt = 3       % Maximum no. of N-R Iterations
-        linear = false
-    end
-    methods
-        function TI = GeneralizedAlpha(varargin)
-            %% Input parsing
-            p = inputParser;
-            addParameter(p,'rho_inf',TI.rho_inf, @(x) isscalar(x) && x<1 && x>0 );
-            addParameter(p,'timestep',TI.h, @(x)validateattributes(x, ...
-                {'numeric'},{'nonempty'}) );
-            addParameter(p,'RelTol',TI.tol, @(x)validateattributes(x, ...
-                {'numeric'},{'nonempty','positive'}) );
-            addParameter(p,'MaxNRit',TI.MaxNRit, @(x)validateattributes(x, ...
-                {'numeric'},{'nonempty','integer','positive'}) );
-            addParameter(p,'linear', TI.linear, @(x)validateattributes(x,{'logical'},{'nonempty'}));
-            addParameter(p,'hmin', TI.hmin, @(x)validateattributes(x, ...
-                {'numeric'},{'nonempty'}) );
-            addParameter(p,'ATS', TI.ATS, @(x)validateattributes(x,{'logical'},{'nonempty'}));
-            
-            parse(p,varargin{:});
-            
-            %% Properties assignment
-            % main time integration parameter that determines numerical
-            % damping of high frequency modes.
-            TI.rho_inf = p.Results.rho_inf;  
-            
-            % from (24) in Arnold & Bruels [1], optimal parameters deduced
-            % by Chung & Hulbert.
-            TI.alpha_m = (2 * TI.rho_inf - 1) / (TI.rho_inf + 1);            
-            TI.alpha_f =  TI.rho_inf / (TI.rho_inf + 1);            
-            % from (20) in Arnold & Bruels [1] to ensure second-order truncation error 
-            TI.gamma = 0.5 + TI.alpha_f - TI.alpha_m;
-            TI.beta = 0.25 * (TI.gamma + 0.5)^2;
-                        
-            % Other parameters by user
-            TI.h = p.Results.timestep;
-            TI.tol = p.Results.RelTol;
-            TI.MaxNRit = p.Results.MaxNRit;
-            TI.hmin = p.Results.hmin;
-            TI.linear = p.Results.linear;
-            
-            % Auxillary parameters (these definition are different from the
-            % ones given in Arnold & Bruels [1]), we solve for acceleration
-            % corrections here instead of displacements.
-            TI.beta_p = TI.h^2 * TI.beta * (1 - TI.alpha_f)/(1-TI.alpha_m);
-            TI.gamma_p = TI.gamma * TI.h * (1 - TI.alpha_f) / (1 - TI.alpha_m);
-        end
-                                
[…]
