### 2022-12-02 #4211 @urban-fasel
@Kevin-Mattheus-Moerman, thanks for the comments! I created the code_ext folder, with the complete distmesh code (this should solve https://github.com/urban-fasel/FlexWingROM/issues/1), cosspace, XFOIL, and yetANotherFECode, and added the license files to code_ext. I replaced two mex functions (crossC and dotC) to simplify the code and I also updated the flexWingROM license to GPLv3.
