### 2022-12-05 #4211 @Kevin-Mattheus-Moerman
@urban-fasel I opened some more issues: https://github.com/urban-fasel/FlexWingROM/issues

Note, it looks like I may need to find an alternative reviewer, as it seems currently your software does not support use on Linux (I use Ubuntu). Can you confirm or advise?
