### 2022-12-05 015fe72
XFOIL precalculated tables for Linux
