+++ example2_NACA6418_unsteadyFSI.m
@@ -8,9 +8,7 @@
 %
 %%%%%%%%%%
 
-clear all
-close all
-clc
+clear; close all; clc
 
 addpath(genpath('code'))
 
@@ -57,7 +55,7 @@ if generateWing
     paramFSI = saveParamFSI(wingDesign,simParam,wingModelStructure,wingModelAero,storeAllData);
 else
 
-    parameters = load(['data/parsim_FSI_', airfoil, '.mat']);
+    parameters = load(['data', filesep, 'parsim_FSI_', airfoil, '.mat']);
     paramFSI = parameters.paramFSI;
 end
 
