+++ MAIN.m
@@ -56,9 +56,7 @@
 %
 %%%%%%%%%%
 
-clear all
-close all
-clc
+clear; close all; clc
 
 addpath(genpath('code'))
 
@@ -114,14 +112,14 @@ if generateWing
     paramFSI = saveParamFSI(wingDesign,simParam,wingModelStructure,wingModelAero,storeAllData);
 else
 %     % load parameters
-%     parameters = load(['data/par_all_FSI_', airfoil, '.mat']);
+%     parameters = load(['data', filesep, 'par_all_FSI_', airfoil, '.mat']);
 %     paramFSI = parameters.paramFSI;
 %     wingDesign = parameters.wingDesign;
 %     simParam = parameters.simParam;
 %     wingModelStructure = parameters.wingModelStructure;
 %     wingModelAero = parameters.wingModelAero;
     
-    parameters = load(['data/parsim_FSI_', airfoil, '.mat']);
+    parameters = load(['data', filesep, 'parsim_FSI_', airfoil, '.mat']);
     paramFSI = parameters.paramFSI;
 end
 
