+++ .gitignore.txt
@@ -1,0 +1,2 @@
+#Ignore MATLAB temporary files
+*.asv
