### 2022-12-05 #4211 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman thanks for your helpful comments! As mentioned directly in the issue, XFOIL is only used for precalculating viscous drag, which I now store in tables for different airfoils that can be loaded if the code is run on Linux. The main code does not use XFOIL and the code should therefore be useful also for Linux users.

Great, it looks like I'll be able to proceed so.
