### 2022-12-09 #4211 @Kevin-Mattheus-Moerman
@urban-fasel I have concluded my review of your software and the paper. I have ticked all boxes but below are some points that need your attention relating to the paper:

- [ ] In the affiliations please spell out USA as `United States of America`.
- [ ] Please add city/town names to all affiliations
- [ ] You may replace `open-source` by `open source` throughout
- [ ] Currently American (e.g. parametrized) and British English (e.g. behaviour) are used. Please choose one to be consistent, and correct accordingly.
