### 2022-12-09 1b15f3d
Create paperV1.md
