+++ docs/paper/paper.md
@@ -14,14 +14,16 @@ authors:
   - name: Andrea Iannelli
     affiliation: 3
   - name: Steven L. Brunton
-    affiliation: 1
+    affiliation: 4
 affiliations:
- - name: Department of Mechanical Engineering, University of Washington, USA
+ - name: Department of Aeronautics, Imperial College, London, United Kingdom
    index: 1
- - name: Department of Aerospace Science and Technology, Politecnico di Milano, Italy
+ - name: Department of Aerospace Science and Technology, Politecnico di Milano, Milan, Italy
    index: 2
- - name: Department of Information Technology and Electrical Engineering, ETH Zurich, Switzerland
+ - name: Department of Information Technology and Electrical Engineering, ETH Zurich, Zurich, Switzerland
    index: 3
+ - name: Department of Mechanical Engineering, University of Washington, Seattle, United States of America
+   index: 4
 date: 27 January 2022
 bibliography: paper.bib
 
@@ -30,7 +32,7 @@ bibliography: paper.bib
 # Summary
 
 Flexible wings pose a considerable modeling challenge, as they involve highly coupled and nonlinear interactions between the aerodynamic and structural dynamics. 
-In this work, we provide an open-source code framework, unifying recent data-driven modeling methods that extend the dynamic mode decomposition with control (DMDc) to model the nonlinear aeroelasticity of flexible wings. 
+In this work, we provide an open source code framework, unifying recent data-driven modeling methods that extend the dynamic mode decomposition with control (DMDc) to model the nonlinear aeroelasticity of flexible wings. 
 Our framework consists of (1) a fully parametrized flexible wing model; (2) a fluid-structure interaction (FSI) solver that couples a detailed finite element model of the wing structure with a 3D unsteady aerodynamic panel method; and (3) three different data-driven reduced-order aeroelastic modeling methods. 
 We demonstrate our framework on two flexible wings and provide tutorials to compare the different data-driven methods. 
 The code is widely applicable and useful for generating accurate and efficient data-driven reduced-order models, and it provides a benchmark for future developments of aeroelastic reduced-order modeling methods.
@@ -38,13 +40,13 @@ The code is widely applicable and useful for generating accurate and efficient d
 # Statement of need
 
 
-Accurate and efficient aeroelastic models are critically important for enabling the optimization and control of highly flexible aero-structures, which are expected to become pervasive in future transportation and energy systems. Indeed, more responsive and deformable aerodynamic surfaces may have transformative impact in efficiency and manoeuvrability, as demonstrated by the incredible performance of biological flight systems [@Combes:2001;@Birch2001nature;@Hedenstrom2007science;@song2008aeromechanics]. These flexible aero-structures pose a considerable modeling challenge, as they involve highly coupled and nonlinear interactions between the aerodynamic and structural dynamics. A range of aeroelastic modeling methods exist, from high-fidelity models able to simulate the coupled fluid flow fields and the deformation of the wing structure [@peskin2002;@Mittal2005;@Goza2017], to control-oriented data-driven models including viscous and unsteady effects [@Brunton2013_JFM;@Hemati2017;@hickner2021data], and lower-fidelity models such as doublet-lattice methods [@Albano1969], strip theory [@Kim2008], and panel methods [@Murua2012;@fasel2021concurrent;@fonzi2020data]. These methods provide accurate models of aeroelastic effects; however, computationally more efficient models are often necessary for optimization and control, even at the expense of some fidelity. This tradeoff between accuracy and efficiency motivates our work on reduced-order models, which model the behaviour of the system with as few states as possible.
[…]
