### 2022-12-09 #4211 @Kevin-Mattheus-Moerman
@urban-fasel thanks for processing those changes. All looks good now. At this point the review is finished and both reviewers have recommended accept. Next, before we can proceed to formal acceptance, please can you work on the following:

- [ ] Please archive a copy of the reviewed software on ZENODO and report back here with the DOI. You can archive a copy manually, but some find these [automated steps](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) useful.

Next ensure that the following ZENODO aspects are correct (you may need to manually edit this:
- [ ] Ensure that the title matches that of the JOSS paper
- [ ] Ensure that the author set and author order matches that of the JOSS paper
- [ ] Ensure that the license listed is the same as for the reviewed software (the default is `other(open)` so check this).
- [ ] We recommend that you add ORCID information for authors

Finally:
- [ ] Please ensure that you have a tagged release on GitHub which is the same as the reviewed software and the same as the archived software on ZENODO.
- [ ] Please report the release/version tag here. Currently the review here lists `v1.0.0`, please let me know if that is updated e.g. as part of the steps shown here. You can alter it as you wish as long as what we assign for this review (I can update it when you tell me to do so) matches a labelled release on your repository as well as the version shown on ZENODO.

Let me know if you have any questions.
