### 2022-12-09 #4211 @urban-fasel
@Kevin-Mattheus-Moerman I archived a copy of the software to ZENODO. The DOI is 10.5281/zenodo.7419465. The tagged release is v1.0.0. Thanks again for your review!
