+++ README.md
@@ -1,5 +1,7 @@
 # FlexWing-ROM
  
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7419465.svg)](https://doi.org/10.5281/zenodo.7419465)
+ 
 The matlab codebase **FlexWing-ROM** models fluid structure interaction (FSI) of flexible wings and generates data-driven reduced order models that can be used for analysis and control.  
 
 The different methods are introduced in the JOSS paper Fasel et al. (2022) -> link will follow.
