### 2022-12-11 #4211 @Kevin-Mattheus-Moerman
@urban-fasel can you check those potentially missing DOI's? ☝️ You can call `@editorialbot check references` to check them
