### 2022-12-12 2587bf7
Update JOSS bib
