+++ docs/paper/paperV1.bib
@@ -8,7 +8,7 @@
   pages={095043},
   year={2017},
   publisher={IOP Publishing},
-  doi = {10.1088/1361-665X/aa7c87}
+  url = {https://doi.org/10.1088/1361-665X/aa7c87}
 }
 
 @article{fasel2021concurrent,
@@ -20,7 +20,7 @@
   pages={1254--1268},
   year={2021},
   publisher={American Institute of Aeronautics and Astronautics},
-  doi = {10.2514/1.J059621}
+  url = {https://doi.org/10.2514/1.J059621}
 }
 
 @article{proctor2016dynamic,
@@ -32,7 +32,7 @@
   pages={142--161},
   year={2016},
   publisher={SIAM},
-  doi = {10.1137/15M1013857}
+  url = {https://doi.org/10.1137/15M1013857}
 }
 
 @article{fonzi2020data,
@@ -44,7 +44,7 @@
   pages={20200079},
   year={2020},
   publisher={The Royal Society Publishing},
-  doi = {10.1098/rspa.2020.0079}
+  url = {https://doi.org/10.1098/rspa.2020.0079}
 }
 
 @article{annoni2017method,
@@ -56,7 +56,7 @@
   pages={582--597},
   year={2017},
   publisher={Wiley Online Library},
-  doi = {10.1002/rnc.3586}
+  url = {https://doi.org/10.1002/rnc.3586}
 }
 
 @article{iannelli2021balanced,
@@ -67,13 +67,13 @@
   pages={106821},
   year={2021},
   publisher={Elsevier},
-  doi = {10.1016/j.ast.2021.106821}
+  url = {https://doi.org/10.1016/j.ast.2021.106821}
 }
 
 @software{jain2020yet,
   author = {Jain, Shobhit and Marconi, Jacopo and Tiso, Paolo},
   title = {Yet{A}nother{FE}code, {Z}enodo},
-  doi = {10.5281/zenodo.4011282},
+  url = {http://doi.org/10.5281/zenodo.4011282},
   version = {1.1},
   year = {2020},
 }
@@ -92,7 +92,7 @@
   volume={13},
   year={2001},
   publisher={Cambridge university press},
-  doi = {10.1017/CBO9780511810329}
+  url = {https://doi.org/10.1017/CBO9780511810329}
 }
 
 @article{newmark1959method,
@@ -104,7 +104,7 @@
   pages={67--94},
   year={1959},
   publisher={American Society of Civil Engineers},
-  doi = {10.1061/JMCEA3.0000098}
+  url = {https://doi.org/10.1061/JMCEA3.0000098}
 }
 
 @article{harder1972interpolation,
@@ -115,7 +115,7 @@
   number={2},
   pages={189--191},
   year={1972},
-  doi = {10.2514/3.44330}
+  url = {https://doi.org/10.2514/3.44330}
 }
 
 @inproceedings{shepard1968two,
@@ -124,7 +124,7 @@
   booktitle={Proceedings of the 1968 23rd ACM national conference},
   pages={517--524},
   year={1968},
-  doi = {10.1145/800186.810616}
+  url = {https://doi.org/10.1145/800186.810616}
 }
 
 @article{molinari2014aero,
@@ -136,7 +136,7 @@
   pages={1940--1951},
   year={2014},
   publisher={American Institute of Aeronautics and Astronautics},
-  doi = {10.2514/1.J052715}
+  url = {https://doi.org/10.2514/1.J052715}
 }
 
 @article{fasel2019reduced,
@@ -148,7 +148,7 @@
   pages={3586--3598},
   year={2019},
   publisher={American Institute of Aeronautics and Astronautics},
-  doi = {10.2514/1.J058019}
+  url = {https://doi.org/10.2514/1.J058019}
 }
[…]
