### 2022-12-12 #4211 @urban-fasel
@Kevin-Mattheus-Moerman I just updated the bib file, all DOIs are ok now.
