### 2022-12-12 #4211 @danielskatz
[The DOI](https://doi.org/10.21105/joss.04211) resolves for me, but [the PDF](https://www.theoj.org/joss-papers/joss.04211/10.21105.joss.04211.pdf) is not available
