### 2022-12-12 #4211 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
